Fix breaking mclda test
In Julia 0.4, one of the characteristic vectors flips sign. I think the answer is still correct, but the test fails because the sign flip is not taken into account.

For now, I've copied the test_approx_eq_vecs from base Julia here to run the test properly. (See JuliaLang/julia#10561 for a PR to move this function into Base.Test)