Fix DimensionMismatch in `loadings` (#229)
When the embedding has a number of zero eigenvalues, `λ` gets truncated.
However, `U` does not, and thus `loadings` throws a `DimensionMismatch`
error on master. This fixes that error.

Alternatively, we could truncate `U` as well, but retaining the full `U`
matrix makes it easier to obtain a full orthonormal basis, if desired.