Add the option to ignore `ConvergenceExceptions` (#233)
The result of a non-converged model might still be interesting. However, since
`ConvergenceExceptions` are thrown these partially fitted models are discarded.

This commit adds a `omit_convergence_exception` argument to all relevant
functions that allows the user to omit such exceptions and return the partially
trained model. The exception is still thrown in the default case to preserve
backwards compatibility.